+++ foundation_headless/tests.lua
@@ -1,4 +1,5 @@
 local mod = assert(foundation_headless)
 
+mod:require("tests/craft_registry_test.lua")
 mod:require("tests/inv_ref_test.lua")
 mod:require("tests/item_stack_test.lua")